Develop

+++ db/migrate/20240602020201_remove_user_id_from_daily_programs.rb
@@ -1,10 +1,13 @@
 class RemoveUserIdFromDailyPrograms < ActiveRecord::Migration[7.0]
   def change
-    # 外部キーが存在する場合のみ削除
-    if foreign_key_exists?(:daily_programs, :users)
-      remove_foreign_key :daily_programs, :users
-    end
+    if table_exists?(:daily_programs)
+      if foreign_key_exists?(:daily_programs, :users)
+        remove_foreign_key :daily_programs, :users
+      end
 
-    remove_column :daily_programs, :user_id, :integer
+      if column_exists?(:daily_programs, :user_id)
+        remove_column :daily_programs, :user_id, :integer
+      end
+    end
   end
 end
